+++ src/events/misc.h
@@ -12,6 +12,7 @@ void	event_invalidUsername(struct irc_message_compo *);
 void	event_local_and_global_users(struct irc_message_compo *);
 void	event_nicknameInUse(struct irc_message_compo *);
 void	event_userModeIs(struct irc_message_compo *);
+void	event_userhost(struct irc_message_compo *);
 void	event_youAreOper(struct irc_message_compo *);
 __SWIRC_END_DECLS